+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### v0.6.0
 
-font-v changes:
+font-v exeecutable changes:
 
 - added head table fontRevision record reporting to report subcommand output (default)
 - added head table fontRevision record write support to write subcommand (default)